+++ crates/wdl-engine/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Fixed
 
+* Validate constraints in the Docker backend, and point error messages to problematic `hints`/`requirements` sections for all backends ([#484](https://github.com/stjude-rust-labs/sprocket/pull/484)).
 * Cache file digests that are calculated during local directory digest
   calculation. Note: this change will bust existing local cache entries ([#488](https://github.com/stjude-rust-labs/sprocket/pull/488)).